voice authenticator/voice authenticator/backend/requirements.txt
fastapi
uvicorn
cryptography
python-jose
pyotp
qrcode
python-dotenv
authlib
pillow
SpeechRecognition
